requirements.txt
python-dotenv==0.21.0
omegaconf==2.2.3
requests==2.28.1
tqdm==4.64.1